update candidate cache
