+++ apps/web/src/pages/api/infinity/candidates-cache.ts
@@ -1,3 +1,4 @@
+import { takeFirstFulfilled } from '@pancakeswap/utils/promise'
 import { POOLS_FAST_REVALIDATE } from 'config/pools'
 import { NextRequest, NextResponse } from 'next/server'
 import { poolQueriesFactory, poolQueryPersistURL } from './edgePoolQueries'
@@ -7,19 +8,37 @@ export const config = {
   runtime: 'edge',
 }
 
+async function fetchCDN(url: string) {
+  const resp = await fetch(url, {
+    method: 'GET',
+    headers: {
+      'Content-Type': 'application/json',
+    },
+  })
+  return resp.json()
+}
+
 export default async function handler(req: NextRequest) {
   const raw = new URL(req.url).search.slice(1)
-  const { chainId, addressA, addressB, protocols } = parseCandidatesQuery(raw)
-
   try {
+    const { chainId, addressA, addressB, protocols } = parseCandidatesQuery(raw)
+
     const poolQueries = poolQueriesFactory(chainId)
+    const query = async () => {
+      const pools = await poolQueries.fetchAllPools(addressA, addressB, chainId, protocols as Protocol[])
+      return pools
+    }
+
     const epoch = Math.floor(Date.now() / POOLS_FAST_REVALIDATE[chainId])
     const { key, cacheKey } = poolQueryPersistURL(addressA, addressB, chainId, protocols as Protocol[], epoch)
-    const pools = await poolQueries.fetchAllPools(addressA, addressB, chainId, protocols as Protocol[])
+    const { url: prev } = poolQueryPersistURL(addressA, addressB, chainId, protocols as Protocol[], epoch - 1)
+    const { result: pools, index } = await takeFirstFulfilled([query(), fetchCDN(prev)])
+
     return responseJson(pools, {
       epoch: Math.floor(Date.now() / POOLS_FAST_REVALIDATE[chainId]),
       url: `${process.env.NEXT_PUBLIC_PROOF_API}/cache/${key}`,
       cacheKey,
+      route: index === 0 ? 'api' : 'cdn',
     })
   } catch (ex) {
     console.error('fetch candidates error', ex)
